reorganization of world context
