+++ src/server/handlers/handlers.ts
@@ -1,4 +1,44 @@
-import { BaseContext, CardContext, GameContext, PenguinContext, RoomContext, SledContext, WorldContext } from "@server/socket-server/world/world";
+import { CardJitsu } from "@server/socket-server/world/card";
+import { SledRace } from "@server/socket-server/world/sled";
+import { PenguinEnvironment, World } from "@server/socket-server/world/world";
+import { WorldGame } from "@server/socket-server/world/world-game";
+import { WorldPenguin } from "@server/socket-server/world/world-penguin";
+import { WorldRoom } from "@server/socket-server/world/world-room";
+import { PenguinMessenger } from "./messenger";
+import { GameData } from "@server/timelines/game-data";
+import { SettingsManager } from "@server/settings";
+import { PenguinRepository } from "@server/database/database";
+import { ClientSocket } from "@server/socket-server/socket-server";
+
+export type PenguinPersister = (p: WorldPenguin, force?: boolean) => void;
+
+type Ctx<Global, AlwaysSingular, EventuallySingular, EventuallyCommon> = Global & AlwaysSingular & ({} | (EventuallySingular & ({} | EventuallyCommon)));
+
+export type WorldContext = Ctx<GlobalContext, AlwaysSingularContext, EventuallySingularContext, PenguinEnvironment>;
+
+export type BaseContext = GlobalContext & AlwaysSingularContext & ({} | EventuallySingularContext);
+export type PenguinContext = BaseContext & EventuallySingularContext & (PenguinEnvironment | {});
+export type RoomContext = PenguinContext & { room: WorldRoom };
+export type GameContext = PenguinContext & { game: WorldGame };
+export type CardContext = PenguinContext & { card: CardJitsu };
+export type SledContext = PenguinContext & { sled: SledRace };
+
+type GlobalContext = {
+  world: World;
+  msg: PenguinMessenger;
+  data: GameData;
+  settings: SettingsManager;
+  db: PenguinRepository;
+  prst: PenguinPersister;
+}
+
+type AlwaysSingularContext = {
+  client: ClientSocket;
+}
+
+type EventuallySingularContext = {
+  penguin: WorldPenguin;
+}
 
 export type HandlerFunction<Ctx extends WorldContext, Args extends readonly any[]> = (ctx: Ctx, ...args: Args) => void | Promise<void>;
